Fix resource start date formatting
