+++ src/Aspire.Dashboard/Components/ResourcesGridColumns/StartTimeColumnDisplay.razor
@@ -1,20 +1,20 @@
 ﻿@using Aspire.Dashboard.Model
 @using Aspire.Dashboard.Resources
 @using System.Globalization
+@using Aspire.Dashboard.Utils
 
-@if (Resource.CreationTimeStamp is not null)
+@if (Resource.CreationTimeStamp is {} timeStamp)
 {
-    var localTime = Resource.CreationTimeStamp.Value.ToLocalTime();
-
-    if (localTime.Date == DateTime.Now.Date)
+    if (timeStamp.ToLocalTime().Date == DateTime.Now.Date)
     {
         // e.g. "08:57:44" (based on user's culture and preferences)
-        @localTime.TimeOfDay.ToString()
+        // Don't include milliseconds as resource server returned time stamp is second precision.
+        @FormatHelpers.FormatTime(timeStamp)
     }
     else
     {
         // e.g. "9/02/2024 08:57:44" (based on user's culture and preferences)
-        @localTime.ToString(CultureInfo.CurrentCulture)
+        @FormatHelpers.FormatDateTime(timeStamp)
     }
 }
